+++ CHANGELOG.md
@@ -4,7 +4,13 @@ History
 1.0.0-alpha2 (unreleased)
 -------------------------
 
-- Nothing changed yet.
+- Creation of feature *Git*
+<br/>
+Currently, there is only one action : git:update-file-umask which update umask for files based on git index. 
+<br/> 
+***Reminder :*** in order to update the umask of a file in git, use command ```git update-index --chmod=+x foo.sh```
+<br/> 
+It can be disabled by setting ```git.auto_umask``` to false in your project ddb.yaml file
 
 
 1.0.0-alpha1 (2020-05-10)