+++ CHANGELOG.md
@@ -8,6 +8,10 @@ History
 based on git index. In order to update permissions of a file in git, use command 
 ```git update-index --chmod=+x foo.sh```. It can be disabled by setting ```git.fix_files_permissions``` to false in 
 configuration.
+- Add `--fail-safe` command line argument to stop on first error.
+- Add `utils.process` module to help running external commands. It makes possible to configure path and additional 
+arguments to any external process invoked by ddb.
+- Default command line argument values can now be customized in configuration using `defaults` key.
 
 
 1.0.0-alpha1 (2020-05-10)